requirements.txt
tweepy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
slack_sdk